Keeping out scenario metric due to errors. Updated inline doc
