+++ VASTTesting/ScenarioMetric.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "VComponent.h"
 #include "VType.h"
+#include "EventTree.h"
 
 /*base class for all metrics in VAST*/
 class ScenarioMetric : public VComponent
@@ -9,144 +10,52 @@ public:
 	/* default constructor*/
 	ScenarioMetric() {};
 
+	/* Constructor to set component data.*/
+	ScenarioMetric(string name, dataMap metricData) 
+	{
+		_myMap = metricData;
+	};
+
 	/* returns the statistical metric value as defined by the extending metric 
 	class.*/
-	virtual VType calculate() { return nullptr; };
-	
-	//declares the virtual function for Vcomponent
-	void update()
+	virtual void calculate() {};
+
+	/* The inherited update function*/
+	virtual void update(timestamp t, dataMap dataMap)
 	{
+		EventTree *temp = getEventTree();
+
+		_myMap["Acceleration"] = dataMap["Acceleration"];
+		_myMap["Velocity"] = dataMap["Speed"];
+		_myMap["Position"] = dataMap["Position"];
+
 		calculate();
-	}
-};
 
-/*
-	calculates the average speed of the chosen vehicle
-*/
-class AverageSpeed : public ScenarioMetric
-{
-public:
-	//default constructor
-	AverageSpeed() {};
+		temp->addEvent(this, t, databaseMap);
+	};
 
-	VType Calculate(float speed)
+	/* Informs the component that the replication is coming to an end.  This function
+	is called by the EventTree to give the component the opportunity to reset data, or
+	processes that must be refreshed in a new replication.
+	bool	another		A signal that there will be another replication
+						after this replication stops.
+	string	runID		The id number used for this replication.*/
+	virtual void stopReplication(bool another, string runID);
+
+	/* Overridden from VComponent.  Returns ID name.*/
+	virtual string getName()
 	{
-		currentAvgSpd += speed;
-		totalSpeeds++;
+		return "Scenario Metric";
+	};
 
-		return currentAvgSpd / totalSpeeds;
-	}
-
-private:
-	Double currentAvgSpd = 0.0;
-	float totalSpeeds = 0;
-};
-
-/*
-	calculates the Maximum Acceleration for the vehicle
-*/
-class MaximumAcceleration : public ScenarioMetric
-{
-public:
-	//default constructor
-	MaximumAcceleration() {};
-
-	VType calculate(Double acceleration)
+	/* Overridden from VComponent.  Returns a copy of the internal data map.*/
+	virtual dataMap getDataMap()
 	{
-		if (acceleration.value() > currentMax.value())
-			currentMax = acceleration;
-
-		return currentMax;
+		// this is an empty map, the default return from ScenarioMetric
+		return _myMap;
 	}
 
-	void printMaximum()
-	{
-		//prints out the maximum value with whatever output type we are using
-	}
-
-private:
-	Double currentMax = 0.0;
-};
-
-/*
-	calculates the Maximum Acceleration for the vehicle
-*/
-class MinimumAcceleration : public ScenarioMetric
-{
-public:
-	//default constructor
-	MinimumAcceleration() {};
-
-	VType calculate(Double acceleration)
-	{
-		if (acceleration.value() < currentMin.value())
-			currentMin = acceleration;
-
-		return currentMin;
-	}
-
-	void printMinimum()
-	{
-		//prints out the minimum value with whatever output type we are using
-	}
-
-private:
-	//set to some unreachable value
-	Double currentMin;
-};
-
-/*
-	calculates the Average Acceleration for the vehicle
-*/
-class AverageAcceleration : public ScenarioMetric
-{
-public:
-	//default constructor
-	AverageAcceleration() {};
-
-	VType calculate(float Speed, int timeStep)
-	{
-		float temp;
-
-		temp = ((previousSpeed - Speed) / timeStep);
-		totalAverages++;
-
-		// these do no inherit the calculate function from the parent 4/25/19 emf
-		/*min.calculate(temp);
-		max.calculate(temp);
-		
-		avgSpeed.calculate(Speed);*/
-		previousSpeed = Speed;
-
-		return currentAverage / timeStep;
-	}
-
-private:
-	// these cannot be instantiated in present condition 4/25/19 emf
-	/*MinimumAcceleration min;
-	MaximumAcceleration max;
-	AverageSpeed avgSpeed;*/
-	Double currentAverage = 0.0;
-	int totalAverages = 0;
-	float previousSpeed = 0;
-};
-
-/*
-	holds the X Y and Z coordinates for the vehicle
-*/
-class XYZCoordinates : public ScenarioMetric
-{
-public:
-
-	XYZCoordinates() {};
-
-	VType calculate(double x, double y, double z)
-	{
-		coordiantes.value(x, y, z);
-
-		return coordiantes;
-	}
-
-private:
-	Vector3 coordiantes;
+	//holds the metric information that is needed
+	dataMap _myMap;
+	dataMap databaseMap;
 };
